Update main.py [ Added text-to-speech feature ]
So after implementing this, if a prompt has been sent by a user sitting inside a VC. The bot will generate the response, prepare a Female TTS voice using gtt library. It will then join the VC where the user is present and will just give the sound output and the user can listen the answer of its question (prompt). After the response is completed the bot will, disconnect and then also send the text like it does.

If the user is not in VC, the bot will work normally like it used to do earlier.

The requirements libraries are already updated in the requirements.txt and imported into the main.py

However this feature is a BETA feature, there are some more things that should be looked and checked before releasing into the public. I use this code in my own bot so I thought to share it as well.